Refactorizing
