+++ llama_ros/llama_ros/langchain/__init__.py
@@ -1,3 +1,4 @@
+from llama_ros.langchain.llama_ros_common import LlamaROSCommon
 from llama_ros.langchain.llama_ros import LlamaROS
 from llama_ros.langchain.chat_llama_ros import ChatLlamaROS
 from llama_ros.langchain.chat_prompt_formatter import ChatPromptFormatter
