+++ llama_ros/llama_ros/llama_client_node.py
@@ -111,12 +111,15 @@ class LlamaClientNode(Node):
         self._spin_thread.start()
 
     def tokenize(self, req: Tokenize.Request) -> Tokenize.Response:
+        self._tokenize_srv_client.wait_for_service()
         return self._tokenize_srv_client.call(req)
 
     def generate_embeddings(self, req: GenerateEmbeddings.Request) -> GenerateEmbeddings.Response:
+        self._embeddings_srv_client.wait_for_service()
         return self._embeddings_srv_client.call(req)
     
     def format_chat_prompt(self, req: ChatMessages.Request) -> ChatMessages.Response:
+        self._format_chat_srv_client.wait_for_service()
         return self._format_chat_srv_client.call(req)
 
     def generate_response(
